+++ source/src/CaloObject/CaloHit.cc
@@ -4,7 +4,7 @@ namespace caloobject
 {
 
 
-  CaloHit::CaloHit(int* cellID, CLHEP::Hep3Vector vec, float energy, int time, CLHEP::Hep3Vector posShift)
+  CaloHit::CaloHit(int* cellID, CLHEP::Hep3Vector vec, float energy, float time, CLHEP::Hep3Vector posShift)
   {
     for(int i=0; i<3; i++)
       _cellID[i]=cellID[i];
